# DefaultObjectPool "devirtualizes" calls to IPooledObjectPolicy