--- a/src/Microsoft.Extensions.ObjectPool/Microsoft.Extensions.ObjectPool.csproj
+++ b/src/Microsoft.Extensions.ObjectPool/Microsoft.Extensions.ObjectPool.csproj
@@ -8,4 +8,8 @@
     <PackageTags>pooling</PackageTags>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="System.Reflection.Emit.Lightweight" Version="4.3.0" />
+  </ItemGroup>
+
 </Project>