--- a/src/Microsoft.Extensions.ObjectPool/Microsoft.Extensions.ObjectPool.csproj
+++ b/src/Microsoft.Extensions.ObjectPool/Microsoft.Extensions.ObjectPool.csproj
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="System.Reflection.Emit.Lightweight" Version="4.3.0" />
+    <PackageReference Include="System.Reflection.Emit.Lightweight" Version="$(ReflectionEmitVersion)" />
   </ItemGroup>
 
 </Project>