@rynowak 2018-02-21
Thanks for trying and sharing this, it's pretty cool. I think we'd definitely look to avoid this level of complexity for something like this. Closing this out in favor of your other PR.